requirements.txt
pyln-client
pyln-testing
pytest
pytest-timeout
pytest-xdist
